refactor: move ll emitter to header
